fix(review): move _CODEX_CAPABILITIES from conftest.py to _helpers.py
conftest.py is for fixtures; module-level constants imported directly
by test files belong in a dedicated helper module.
