+++ tests/workspace/test_session_skills_provider.py
@@ -18,7 +18,7 @@ from autoskillit.workspace.session_skills import (
     SkillsDirectoryProvider,
     resolve_ephemeral_root,
 )
-from tests.workspace.conftest import _CODEX_CAPABILITIES
+from tests.workspace._helpers import _CODEX_CAPABILITIES
 
 pytestmark = [pytest.mark.layer("workspace"), pytest.mark.small]
